Refactor yes/no authorization
